stack linked repos
